moved the game checks to their own file
